Check dataset updates (#5)
* feat: added vertical datum check in pc

Created vertical datum check to identify if georeferencing wkt is incorrect. Looks for NZVD2016 or New Zealand Vertical Datum 2016 in the wkt of the lasinfo output.

Still to do is add a attribute in pc tile index for which tiles have incorrect wkt reference for vertical datum.

* feat: updated point data format to include more

Originally only checked for point data format 6 but now includes 7, 8, 9, and 10.

* updated where statements

Old where dimensions are now deprecated.